+++ apps/node/lodge/src/store/game/index.ts
@@ -1,5 +1,4 @@
 export * from "./cards"
 export * from "./card-values"
 export * from "./card-groups"
-
-export * from "./game-tx"
+export * from "./transactions"
